js: Add the callback function for submit event
This will be invoked when we get a submit event. When we do receive
a submit event, we retrieve the form value from that event and then
run the Array filter method on each element of our fetchedData Array
to filter in the ones that contain the tag value passed in the form.